Move custom graphics items to `src/gui`.
